Rigidbody revision (2)
- Dash를 일정한 속도로 계속 가는게 아니라 DashFrictionCoef를 지정해서 점점 감속하는 형식으로 수정함

+++ CTextBoxUI.cpp
@@ -112,7 +112,7 @@ void CTextBoxUI::Render_DerivedUI()
 		int cursorX = (int)vTextPos.x + textSize.cx;
 		int cursorY = (int)vTextPos.y;
 		MoveToEx(BackDC, cursorX, cursorY, NULL);
-		LineTo(BackDC, cursorX, cursorY + vScale.y - 6.f);
+		LineTo(BackDC, cursorX, (int)(cursorY + vScale.y - 6.f));
 	}
 
 }
